bug fixed
this bug is:
when particle do not hit the coil, induction voltages and  sample times will all be zero, but when doing fft on it, it will lead to None result.
i fixed this bug by adding a default sample step when sample step is zero.
results shows that there will not be None in output Dict